Issue #285 fixed

+++ packages/Webkul/Admin/src/Resources/views/leads/view/activity-list.blade.php
@@ -302,6 +302,10 @@
                 },
 
                 remove: function(activity) {
+                    if (! confirm('Do you really want to perform this action?')) {
+                        return;
+                    }
+
                     var self = this;
 
                     this.$http.delete("{{ route('admin.activities.delete') }}/" + activity['id'])
